Let the recipient cancel
So, Apple stated that the recipient of the exchange should be able to cancel the exchange. My testing however, indicates that the recipient can only reply or let the exchange time out. There is no cancel by the recipient.

However, I make this dead end available to the Apple engineer.