Requiring iOS 13
There is a bug in iOS 12 relating to encoding string JSON fragments. So to make it easy for myself, I just bumped the required iOS version number. Sorry iOS 12 users, it’s annoying - but .. yeah. I might fix this by using an array instead of a string for the match data.